Fix Third Component
removed divs. added hover color to buttons. My "buttons" are actually
links. I just gave them a class of button. So they perform as links not
buttons. played with positioning a bit. Still need help to get
everything to line up properly. I think the problem is the original was
so un-uniform. I'm not sure how to do this with out putting things in
rows and columns. So, I put it into a table!